minor changes - lib

+++ lib/helpers.ts
@@ -1,10 +1,11 @@
 import axios from "axios";
-import { PDFDocument } from 'pdf-lib';
+import { PDFDocument } from "pdf-lib";
 import { writeFile } from "fs";
-import * as fs from 'fs';
-import path from 'path';
+import * as fs from "fs";
+import path from "path";
 import { Document } from "langchain/document";
 import { RecursiveCharacterTextSplitter } from "langchain/text_splitter";
+import { UnstructuredLoader } from "langchain/document_loaders/fs/unstructured";
 import { PDFLoader } from "langchain/document_loaders/fs/pdf";
 
 /**
@@ -13,16 +14,20 @@ import { PDFLoader } from "langchain/document_loaders/fs/pdf";
  * @param {Object} url - The URL of the PDF file to load.
  * @return {Promise<Buffer>} A promise that resolves with the loaded PDF file as a Buffer.
  */
-export async function loadPdfFromUrl({ url }: { url: string; }): Promise<Buffer> {
-	try {
-		console.log("Loading PDF from URL...");
-		const response = await axios.get(url, { responseType: "arraybuffer", });
-		console.log("PDF loaded successfully 📜📜📜");
-		return response.data;
-	} catch (error) {
-		console.log(error);
-		throw new Error("PDF fetching failed 🩻");
-	}
+export async function loadPdfFromUrl({
+  url,
+}: {
+  url: string;
+}): Promise<Buffer> {
+  try {
+    console.log("Loading PDF from URL...");
+    const response = await axios.get(url, { responseType: "arraybuffer" });
+    console.log("PDF loaded successfully 📜📜📜");
+    return response.data;
+  } catch (error) {
+    console.log(error);
+    throw new Error("PDF fetching failed 🩻");
+  }
 }
 
 /**
@@ -32,7 +37,10 @@ export async function loadPdfFromUrl({ url }: { url: string; }): Promise<Buffer>
  * @param {Array<number>} pagesToDelete - An array of page numbers to delete.
  * @return {Promise<Buffer>} - The modified PDF document as a Buffer.
  */
-export async function deletePages(pdf: Buffer, pagesToDelete: Array<number>): Promise<Buffer> {
+export async function deletePages(
+  pdf: Buffer,
+  pagesToDelete: Array<number>,
+): Promise<Buffer> {
   try {
     const pdfDoc = await PDFDocument.load(pdf);
 
@@ -41,7 +49,7 @@ export async function deletePages(pdf: Buffer, pagesToDelete: Array<number>): Pr
       pdfDoc.removePage(pageNumber - numToOffsetBy);
       numToOffsetBy += 1;
     }
-    console.log('Pages deleted successfully 🗑️');
+    console.log("Pages deleted successfully 🗑️");
     const pdfBytes = await pdfDoc.save();
     return Buffer.from(pdfBytes);
   } catch (error) {
@@ -50,33 +58,42 @@ export async function deletePages(pdf: Buffer, pagesToDelete: Array<number>): Pr
   }
 }
 
-export async function convertPdfToDocuments(pdf: Buffer, pdfTitle: string): Promise<Document[]> {
+export async function convertPdfToDocuments(
+  pdf: Buffer,
+  pdfTitle: string,
+): Promise<Document[]> {
+  if (!process.env.UNSTRUCTURED_API_KEY) throw new Error();
+
+  // create file name for each Document created from the PDF using UUIDv4
+  const fileName = pdfTitle.toLowerCase().replace(/ /g, "-");
+
+  // check if folder exists
+  if (!fs.existsSync("pdf")) fs.mkdirSync("pdfs");
+
+  const pdfPath = `pdfs/${fileName}.pdf`;
+  // Save Document to file
+  fs.writeFileSync(pdfPath, pdf, "binary");
+
+  // const loader = new PDFLoader(pdfPath);
+  // Use with PDFLoader
+  // const splitter = new RecursiveCharacterTextSplitter({
+  //   chunkSize: 1000,
+  //   chunkOverlap: 200,
+  // });
+  // const docs = await loader.loadAndSplit(splitter);
+
+  // create Unstructured loader if Unstructured API key is provided
+  const loader = new UnstructuredLoader(pdfPath, {
+    apiKey: process.env.UNSTRUCTURED_API_KEY,
+    apiUrl: process.env.UNSTRUCTURED_NGROK_DIRECT_URL,
+    strategy: "hi_res",
+  });
+
   try {
-    // create file name for each Document created from the PDF using UUIDv4
-    const fileName = pdfTitle.toLowerCase().replace(/ /g, "-");
-    const pdfPath = `pdfs/${fileName}.pdf`;
-    // Save Document to file
-    fs.writeFileSync(pdfPath, pdf, "binary");
-
-    const loader = new PDFLoader(pdfPath);
-    /* // create Unstructured loader if Unstructured API key is provided
-    const loader = new UnstructuredLoader(pdfPath, {
-      apiKey: process.env.UNSTRUCTURED_API_KEY,
-      strategy: "hi_res",
-    });
-    */
-
-    const splitter = new RecursiveCharacterTextSplitter({
-      chunkSize: 1000,
-      chunkOverlap: 200,
-    });
-    const docs = await loader.loadAndSplit(splitter);
-
-    // delete temporary PDF file 
+    const docs = await loader.load();
+    // delete temporary PDF file
     fs.unlinkSync(pdfPath);
-
     console.log("PDF converted successfully 🔀🔀");
-
     return docs;
   } catch (error) {
     console.log(error);
